Re-organize to move the app-like stuff to conftest.
